+++ src/event-handlers.js
@@ -1,0 +1,29 @@
+import { renderTodos, clearNewTodoInput, getTodoId } from "./ui";
+import { getAllTodos, removeTodo, updateTodo, addTodo } from "./data";
+
+export function onLoadEventHandler() {
+    renderTodos(getAllTodos())
+}
+
+export function newTodoEventHandler(event) {
+    let text = event.target.value
+    addTodo({
+        id: Date.now(),
+        text: text,
+        completed: false
+    })
+    renderTodos(getAllTodos())
+    clearNewTodoInput()
+}
+
+export function removeTodoEventHandler(event) {
+    const id = getTodoId(event.target)
+    removeTodo(id)
+    renderTodos(getAllTodos())
+}
+export function toggleTodoEventListener(event) {
+    const id = getTodoId(event.target)
+    const isCompleted = event.target.checked
+    updateTodo(id, isCompleted)
+    renderTodos(getAllTodos())
+}
