+++ src/todos.json
@@ -1,0 +1,12 @@
+[
+    {
+        "id": 1560865205317,
+        "text": "Buy eggs",
+        "completed": false
+    },
+    {
+        "id": 1560865205318,
+        "text": "Do 15 min exercise",
+        "completed": false
+    }
+]
